modify review,add containerdInstaller.go file,add error return
